feat. 뉴스 기능 추가
